Output dot/graphviz
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@2082 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/fst2txt-functionality.sh
@@ -7,6 +7,12 @@ for i in "" .sfst .ofst .foma; do
         if ! diff test.txt $srcdir/cat.txt > /dev/null 2>&1 ; then
             exit 1
         fi
+        if ! ../../tools/src/hfst-fst2txt -f dot < cat.hfst$i > test.txt ; then
+            exit 1
+        fi
+        if ! dot test.txt > /dev/null 2>&1 ; then
+            exit 1
+        fi
         rm test.txt;
     fi
 done
